==> 4_Kaggle_Submision.ipynb <==
# Kaggle submission

## Import needed libraries

import pandas as pd
import kaggle

## Load prediction

Choose the csv file you want to submit to Kaggle.  
They are stored in 'results' directory.

kaggle_submission = pd.read_csv("./results/kaggle_sub_XGBClassifier.csv")
kaggle_submission

## Submit to Kaggle

The first argument is the path to the csv file containing your submission.  
The second argument is the comment to join to the submission.  
The third argument is the name of the Kaggle competition you are submitting to.

kaggle.api.competition_submit("./results/kaggle_sub_XGBClassifier.csv", "XGBClassifier", "GiveMeSomeCredit")
